♻️(persons) move persons to the courses app
There was not reason to have a separate app for persons and it
is easier if all models are in the same app.
