+++ src/richie/apps/courses/models/__init__.py
@@ -6,3 +6,4 @@ from .blog import *
 from .category import *
 from .course import *
 from .organization import *
+from .person import *
