+++ src/richie/apps/courses/templates/courses/cms/_person_detail.scss
