+++ src/richie/apps/courses/templates/courses/plugins/_person_plugin.scss
