+++ src/richie/apps/courses/templates/courses/plugins/person.html
