+++ tests/apps/courses/utils.py
